Organize test into multiple files
