+++ tests/conftest.py
@@ -1,0 +1,11 @@
+import pytest
+
+from app.app import app
+
+
+@pytest.fixture
+def client():
+    app.config["TESTING"] = True
+
+    with app.test_client() as client:
+        yield client
